Struct accountable_refcell::RefCell

pub struct RefCell<T: ?Sized> {
    pub(crate) borrows: RefCell<BorrowData>,
    pub(crate) inner: RefCell<T>,
}

A RefCell that tracks outstanding borrows and reports stack traces for dynamic borrow failures.

Fields
borrows: RefCell<BorrowData>inner: RefCell<T>
Implementations

impl<T> RefCell<T>

pub fn new(value: T) -> RefCell<T>

Create a new RefCell value.

pub fn into_inner(self) -> T

Discard this RefCell and return the value stored inside of it.

impl<T: ?Sized> RefCell<T>

pub fn borrow(&self) -> Ref<'_, T>

Borrow the value stored in this cell immutably. Panics if any outstanding mutable borrows of the same cell exist.

pub fn try_borrow(&self) -> Result<Ref<'_, T>, BorrowError>

pub fn borrow_mut(&self) -> RefMut<'_, T>

Borrow the value stored in this cell mutably. Panics if there are any other outstanding borrows of this cell (mutable borrows are unique, i.e. there can only be one).

pub fn try_borrow_mut(&self) -> Result<RefMut<'_, T>, BorrowMutError>

pub fn as_ptr(&self) -> *mut T

pub unsafe fn try_borrow_unguarded(&self) -> Result<&T, BorrowError>

impl<T> RefCell<T>

pub fn replace(&self, t: T) -> T

Corresponds to https://doc.rust-lang.org/std/cell/struct.RefCell.html#method.replace.

pub fn replace_with<F: FnOnce(&mut T) -> T>(&self, f: F) -> T

Corresponds to https://doc.rust-lang.org/std/cell/struct.RefCell.html#method.replace_with.

impl<T: Default> RefCell<T>

pub fn take(&self) -> T

Corresponds to https://doc.rust-lang.org/std/cell/struct.RefCell.html#method.take.