Crate accountable_refcell

Structs

BorrowData

BorrowRecord

Ref

An immutable reference to the value stored in a RefCell.

RefBorrowData

RefCell

A RefCell that tracks outstanding borrows and reports stack traces for dynamic borrow failures.

RefMut

A mutable reference to the value stored in the associated RefCell.

Functions

print_filtered_backtrace

Print a backtrace without any frames from the backtrace library.

ref_filter_map

ref_mut_filter_map